+++ lib/globals.dart
@@ -20,8 +20,8 @@ final GlobalKey<ScaffoldMessengerState> scaffoldMessengerKey =
 final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
 const String prgName = 'easysend';
-const String progVersion = '0.2.260812';
-const int buildNumber = 82;
+const String progVersion = '0.2.260813';
+const int buildNumber = 83;
 const String progAuthor = 'Eugen';
 
 const String langFile = 'assets/locales.json';